feat(code):代码优化; feat(UI):锁住窗体大小

+++ OpenProtocolInterpreter/Curve/Mid7410.cs
@@ -129,8 +129,8 @@ namespace OpenProtocolInterpreter.Curve
                 m7410.ParameterSetId = int.Parse(Encoding.ASCII.GetString(dataArray.Skip(6).Take(3).ToArray()));   //pset号
                 if (Revision == 2)
                 {
-                    m7410.TorqueCoefficient = float.Parse(Encoding.ASCII.GetString(dataArray.Skip(39).Take(14).ToArray()));  //扭矩系数
-                    m7410.AngleCoefficient = float.Parse(Encoding.ASCII.GetString(dataArray.Skip(55).Take(14).ToArray()));   //角度系数
+                    m7410.TorqueCoefficient = double.Parse(Encoding.ASCII.GetString(dataArray.Skip(39).Take(14).ToArray()));  //扭矩系数
+                    m7410.AngleCoefficient = double.Parse(Encoding.ASCII.GetString(dataArray.Skip(55).Take(14).ToArray()));   //角度系数
                     m7410.NumMeasurementPoints = int.Parse(Encoding.ASCII.GetString(dataArray.Skip(87).Take(4).ToArray()));   //点数
                     m7410.NumSegments = int.Parse(Encoding.ASCII.GetString(dataArray.Skip(93).Take(2).ToArray()));  //分了几帧
                     m7410.SegmentID = int.Parse(Encoding.ASCII.GetString(dataArray.Skip(97).Take(2).ToArray()));  //当前帧数
@@ -139,8 +139,8 @@ namespace OpenProtocolInterpreter.Curve
                 }
                 else
                 {
-                    m7410.TorqueCoefficient = float.Parse(Encoding.ASCII.GetString(dataArray.Skip(27).Take(14).ToArray()));  //扭矩系数
-                    m7410.AngleCoefficient = float.Parse(Encoding.ASCII.GetString(dataArray.Skip(43).Take(14).ToArray()));   //角度系数
+                    m7410.TorqueCoefficient = double.Parse(Encoding.ASCII.GetString(dataArray.Skip(27).Take(14).ToArray()));  //扭矩系数
+                    m7410.AngleCoefficient = double.Parse(Encoding.ASCII.GetString(dataArray.Skip(43).Take(14).ToArray()));   //角度系数
                     m7410.NumMeasurementPoints = int.Parse(Encoding.ASCII.GetString(dataArray.Skip(59).Take(4).ToArray()));   //点数
                     m7410.NumSegments = int.Parse(Encoding.ASCII.GetString(dataArray.Skip(65).Take(2).ToArray()));  //分了几帧
                     m7410.SegmentID = int.Parse(Encoding.ASCII.GetString(dataArray.Skip(69).Take(2).ToArray()));  //当前帧数
